refactor(client): move gameObject from types to objects
