+++ client/src/scripts/objects/gameObject.ts
